requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
dnspython==2.7.0
idna==3.10
python-nmap==0.7.1
requests==2.32.3
urllib3==2.4.0
whois==1.20240129.2
